extract LDFlagListener from client to common bindings
