+++ libs/client-sdk/include/launchdarkly/client_side/bindings/c/sdk.h
@@ -9,6 +9,7 @@
 #include <launchdarkly/bindings/c/context.h>
 #include <launchdarkly/bindings/c/data/evaluation_detail.h>
 #include <launchdarkly/bindings/c/export.h>
+#include <launchdarkly/bindings/c/flag_listener.h>
 #include <launchdarkly/bindings/c/listener_connection.h>
 #include <launchdarkly/bindings/c/memory_routines.h>
 #include <launchdarkly/bindings/c/status.h>
@@ -411,52 +412,6 @@ LDClientSDK_AllFlags(LDClientSDK sdk);
  * @param sdk SDK.
  */
 LD_EXPORT(void) LDClientSDK_Free(LDClientSDK sdk);
-
-typedef void (*FlagChangedCallbackFn)(char const* flag_key,
-                                      LDValue new_value,
-                                      LDValue old_value,
-                                      bool deleted,
-                                      void* user_data);
-
-/**
- * Defines a feature flag listener which may be used to listen for flag changes.
- * The struct should be initialized using LDFlagListener_Init before use.
- */
-struct LDFlagListener {
-    /**
-     * Callback function which is invoked for flag changes.
-     *
-     * The provided pointers are only valid for the duration of the function
-     * call (excluding UserData, whose lifetime is controlled by the caller).
-     *
-     * @param flag_key The name of the flag that changed.
-     * @param new_value The new value of the flag. If there was not an new
-     * value, because the flag was deleted, then the LDValue will be of a null
-     * type. Check the deleted parameter to see if a flag was deleted.
-     * @param old_value The old value of the flag. If there was not an old
-     * value, for instance a newly created flag, then the Value will be of a
-     * null type.
-     * @param deleted True if the flag has been deleted.
-     */
-    FlagChangedCallbackFn FlagChanged;
-
-    /**
-     * UserData is forwarded into callback functions.
-     */
-    void* UserData;
-};
-
-/**
- * Initializes a flag listener. Must be called before passing the listener
- * to LDClientSDK_FlagNotifier_OnFlagChange.
- *
- * Create the struct, initialize the struct, set the FlagChanged handler
- * and optionally UserData, and then pass the struct to
- * LDClientSDK_FlagNotifier_OnFlagChange.
- *
- * @param listener Listener to initialize.
- */
-LD_EXPORT(void) LDFlagListener_Init(struct LDFlagListener listener);
 
 /**
  * Listen for changes for the specific flag.
